enable to build without Oracle JDBC driver
