+++ embulk-output-oracle/src/test/resources/yml/test-replace-long-name.yml
@@ -1,0 +1,25 @@
+in:
+  type: file
+  path_prefix: '/data/test1/test1.csv'
+  parser:
+    charset: UTF-8
+    newline: CRLF
+    type: csv
+    delimiter: ','
+    quote: ''
+    columns:
+    - {name: id, type: string}
+    - {name: varchar2_item, type: string}
+    - {name: integer_item, type: long}
+    - {name: number_item, type: string}
+    - {name: date_item, type: timestamp, format: '%Y/%m/%d'}
+    - {name: timestamp_item, type: timestamp, format: '%Y/%m/%d %H:%M:%S'}
+out:
+    type: oracle
+    host: localhost
+    database: TESTDB
+    user: TEST_USER
+    password: test_pw
+    table: TEST12345678901234567890123456
+    mode: replace
+    #driver_path: driver/ojdbc7.jar
